pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gcsfs>=2026.3.0",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=3.10.1",
    "kafka-python>=2.3.1",
    "pandas>=3.0.1",
    "pgcli>=4.4.0",
    "psycopg[binary,pool]>=3.3.3",
    "pyarrow>=23.0.1",
    "sqlalchemy>=2.0.48",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pgcli>=4.4.0",
]
